requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
groq==0.13.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
streamlit==1.39.0
httpx==0.27.2
pandas==2.2.3
python-multipart==0.0.12
